scripts: zfs-root-generator: avoid loading zfs.ko when checking pools
When checking if the target pool is imported to determine whether to
parse the nested datasets, avoid using `zpool` and `zfs` tools to
prevent zfs.ko from being auto-loaded. Instead, check if the target
pool is imported by testing for its directory in `/proc/spl/kstat/zfs`.
If ZFS has not yet been loaded, the entire hierarchy will not exist,
which is logically equivalent to the pool not being imported.

This is desirable because generators execute fully synchronously and
we do not want to delay the boot process by the time it takes zfs.ko
to load when the generator is first executed (and the pool would not
have been imported anyway).