scripts: zfs-root-generator: document sysroot-usr.mount divergence
There is `initrd-usr-fs.target`, which at face value should be used for
the `/usr` mount. However, there is a subtle divergence to the normal
fstab-powered bootup: systemd mounts it into `/sysusr/usr`, and _that_
is what is serialized by `initrd-usr-fs.target`. It is then bind-mounted
into `/sysroot/usr` and that mount is added to `initrd-fs.target`. All
of this is reinforced by the fact that `initrd-usr-fs.target` is not
ordered against `initrd-switch-root.target`, which means that naïvely
using it to mount `/sysroot/usr` leads to breakage.

The only place where `initrd-usr-fs.target` is used non-trivially is in
`systemd-repart.service`, which is supposed to create a volatile rootfs
based on the information in `/usr` (presumably mounted to the staging
point described above, as rootfs would not exist by that time). We,
however, do not support a volatile rootfs under a ZFS-backed `/usr`,
so do not use that target.