scripts: zfs-prepare-rootfs: fix race when nested datasets are in use
When nested datasets are in use, `initrd-fs.target` is enqueued earlier
than we have a chance to generate the units for the nested datasets and
add them to the dependencies of `initrd-fs.target`. This means that the
newly generated units are not part of the initial transaction and are
not well-ordered against the root switch.

For lack of a better way to fix this, restart `initrd-fs.target` after
generating the units to reenqueue the target will its new dependencies.